requirements.txt
torch
torchvision
numpy
pandas
pillow
scikit-learn
tqdm
pyyaml
matplotlib

huggingface-hub
python-dotenv
